refactor: split home urls into shared and per-site (#1598)
