--- app/components/Home/common/constants.ts
@@ -1,12 +1,0 @@
-export const URLS = {
-  ACCESS_CONSORTIUM: "https://access-ci.org",
-  EARTH_BIOGENOME_PROJECT: "https://www.earthbiogenome.org",
-  GALAXY_SYSTEM: "https://brc.usegalaxy.org",
-  GALAXY_SYSTEM_VGP: "https://vgp.usegalaxy.org",
-  JETSTREAM2: "https://jetstream-cloud.org",
-  NCBI_DATASETS: "https://www.ncbi.nlm.nih.gov/datasets",
-  T2T_CONSORTIUM: "https://sites.google.com/ucsc.edu/t2tworkinggroup/home",
-  TACC: "https://tacc.utexas.edu",
-  UCSC_GENOME_BROWSER: "https://genome.ucsc.edu",
-  VERTEBRATE_GENOMES_PROJECT: "https://vertebrategenomesproject.org",
-};
