+++ packages/shared/views/HomeView/urls.ts
@@ -1,0 +1,5 @@
+export const SHARED_URLS = {
+  NCBI_DATASETS: "https://www.ncbi.nlm.nih.gov/datasets",
+  TACC: "https://tacc.utexas.edu",
+  UCSC_GENOME_BROWSER: "https://genome.ucsc.edu",
+};
