+++ sites/brc-analytics/views/HomeView/urls.ts
@@ -1,0 +1,6 @@
+import { SHARED_URLS } from "@repo/shared/views/HomeView/urls";
+
+export const URLS = {
+  ...SHARED_URLS,
+  GALAXY_SYSTEM: "https://brc.usegalaxy.org",
+};
