+++ sites/ga2/views/HomeView/urls.ts
@@ -1,0 +1,9 @@
+import { SHARED_URLS } from "@repo/shared/views/HomeView/urls";
+
+export const URLS = {
+  ...SHARED_URLS,
+  EARTH_BIOGENOME_PROJECT: "https://www.earthbiogenome.org",
+  GALAXY_SYSTEM_VGP: "https://vgp.usegalaxy.org",
+  JETSTREAM2: "https://jetstream-cloud.org",
+  VERTEBRATE_GENOMES_PROJECT: "https://vertebrategenomesproject.org",
+};
